Refactoring validation
