+++ src/components/BlenderLogbook/validation.ts
@@ -1,7 +1,5 @@
 import * as yup from 'yup';
-import { FIELD_REQUIRED } from '../UserSettings/validation';
-
-const FIELD_NUMBER = 'Annetun arvon on oltava numero';
+import { FIELD_NUMBER, FIELD_REQUIRED } from '../../lib/validationUtils';
 
 export const BLENDER_FILLING_EVENT_VALIDATION_SCHEMA = yup.object().shape({
   additionalInformation: yup.string().optional(),
