+++ src/components/Logbook/validation.ts
@@ -1,5 +1,5 @@
 import * as yup from 'yup';
-import { FIELD_REQUIRED } from '../UserSettings/validation';
+import { FIELD_REQUIRED } from '../../lib/validationUtils';
 
 export const AIR_FILLING_EVENT_VALIDATION_SCHEMA = yup.object().shape({
   additionalInformation: yup.string().optional(),
